Stochastic GM+E Restarts v2 (#376)
Updates the NUOPC cap to write the stochastic physics package's restart file in the run directory whenever using the Stochastic GM+E parameterization.

When restarting, enables the stochastic physics package to read a restart file from a name/location other than the default. If the new MOM6 namelist parameter CESM_RESTFILE is True then it uses the mom6 restart filename but with .r_stoch. in place of .r.. If CESM_RESTFILE is False (default) it simply uses whatever the default restart file name is in the stochastic physics package. (Edit: This PR no longer uses the CESM_RESTFILE parameter. It just reads the restart file directly from the run directory.) A separate PR to ESCOMP/stochastic_physics updates that package to allow the restart file name to change from its default value, which was originally hard-coded.

This partially addresses #360. Changes will be needed in MOM_interface to make the whole thing work; specifically, we should always set CESM_RESTFILE = True in user_nl_mom, and when CONTINUE_RUN is true we need to modify the stochini parameter in the &nam_stochy section of input.nml to take the value .true..

This PR takes an alternate approach to #373, which required the restart file name to be set in the stochastic physics namelist. I've closed #373 in preference to this approach.